+++ app/upgrades.go
@@ -11,6 +11,7 @@ import (
 	"github.com/NibiruChain/nibiru/app/upgrades/v1_0_1"
 	"github.com/NibiruChain/nibiru/app/upgrades/v1_0_2"
 	"github.com/NibiruChain/nibiru/app/upgrades/v1_1_0"
+	"github.com/NibiruChain/nibiru/app/upgrades/v1_2_0"
 )
 
 var Upgrades = []upgrades.Upgrade{
@@ -18,6 +19,7 @@ var Upgrades = []upgrades.Upgrade{
 	v1_0_2.Upgrade,
 	v1_0_3.Upgrade,
 	v1_1_0.Upgrade,
+	v1_2_0.Upgrade,
 }
 
 func (app *NibiruApp) setupUpgrades() {